changed type to int and Index

+++ include/CORA/CORA_problem.h
@@ -67,11 +67,11 @@ struct Manifolds {
 class Problem {
 private:
   /** dimension of the pose and landmark variables e.g., SO(dim_) */
-  const int64_t dim_;
+  const int dim_;
 
   /** rank of the relaxation e.g., the latent embedding space of Stiefel
    * manifold */
-  int64_t relaxation_rank_;
+  int relaxation_rank_;
 
   // maps from pose symbol to pose index (e.g., x1 -> 0, x2 -> 1, etc.)
   std::map<Symbol, int> pose_symbol_idxs_;
@@ -178,7 +178,7 @@ private:
   Index getTranslationIdx(const Symbol &trans_symbol) const;
 
 public:
-  Problem(int64_t dim, int64_t relaxation_rank,
+  Problem(int dim, int relaxation_rank,
           Formulation formulation = Formulation::Explicit,
           Preconditioner preconditioner = Preconditioner::BlockCholesky)
       : dim_(dim),
@@ -234,11 +234,15 @@ public:
   size_t getDataMatrixSize() const;
 
   Formulation getFormulation() const { return formulation_; }
-  size_t dim() const { return dim_; }
-  size_t numPoses() const { return pose_symbol_idxs_.size(); }
-  size_t numLandmarks() const { return landmark_symbol_idxs_.size(); }
-  size_t numRangeMeasurements() const { return range_measurements_.size(); }
-  size_t numTranslationalStates() const { return numPoses() + numLandmarks(); }
+  int dim() const { return dim_; }
+  int numPoses() const { return static_cast<int>(pose_symbol_idxs_.size()); }
+  int numLandmarks() const {
+    return static_cast<int>(landmark_symbol_idxs_.size());
+  }
+  int numRangeMeasurements() const {
+    return static_cast<int>(range_measurements_.size());
+  }
+  int numTranslationalStates() const { return numPoses() + numLandmarks(); }
 
   /*****  Riemannian optimization functions  *******/
